Reorganize test suites
